\name{nome_funzione}
\alias{nome_funzione}
\title{Titolo conciso della funzione}
\description{
Descrizione estesa della funzione.
}
\usage{
nome_funzione(arg1, arg2, ...)
}
\arguments{
  \item{arg1}{Descrizione di arg1.}
  \item{arg2}{Descrizione di arg2.}
}
\details{
Dettagli aggiuntivi sul funzionamento interno, assunzioni, note tecniche.
}
\value{
Descrizione dell’output restituito dalla funzione.
}
\examples{
# Esempio minimo
ris <- nome_funzione(...)
}
